Monster 필드 Intgeger -> Float 으로 변경

+++ ooptest/src/main/java/com/biginsight/ooptest/domain/Monster.java
@@ -20,9 +20,9 @@ public class Monster {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer hp;
-    private Integer attackPower; // 공격력
-    private Integer defensePower; // 방어력
-    private Integer counterattackRate; // 반격 (%) -> 캐릭터의 '회피율' 과 같은 개념으로 생각하고 일단 Skill 로 가져가지 않기로 함.
+    private Float hp;
+    private Float attackPower; // 공격력
+    private Float defensePower; // 방어력
+    private Float counterattackRate; // 반격 (%) -> 캐릭터의 '회피율' 과 같은 개념으로 생각하고 일단 Skill 로 가져가지 않기로 함.
 
 }
